MQTT/requirements.txt
psycopg2
configparser
web3
paho-mqtt
python-etcd
